requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
pymupdf>=1.24.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.0
